+++ sdmx/writer/xml.py
@@ -1,4 +1,4 @@
-"""SDMXML v2.1 writer."""
+"""SDMX-ML v2.1 writer."""
 # Contents of this file are organized in the order:
 #
 # - Utility methods and global variables.
@@ -7,7 +7,7 @@
 
 import logging
 from pathlib import Path
-from typing import Iterable, IO, List, Optional, cast
+from typing import Iterable, IO, Literal, List, Optional, cast
 
 from lxml import etree
 from lxml.builder import ElementMaker
@@ -46,109 +46,9 @@ def to_xml(obj, **kwargs):
     NotImplementedError
         If writing specific objects to SDMX-ML has not been implemented in :mod:`sdmx`.
     """
+    kwargs.setdefault("encoding", "utf-8")
+    kwargs.setdefault("xml_declaration", True)
     return etree.tostring(writer.recurse(obj), **kwargs)
-
-
-def validate_xml(msg: Path | IO, schema_dir: Optional[Path] = None) -> bool:
-    """Validate and SDMX message against the XML Schema (XSD) documents.
-
-    The XML Schemas must first be installed or validation will fail. See
-    :func:`sdmx.install_schemas` to download the schema files.
-
-    Parameters
-    ----------
-    msg
-        A SDMX-ML Message formatted XML file.
-    schema_dir
-        The directory to XSD schemas used to validate the message.
-
-    Returns
-    -------
-    bool
-        True if validation passed. False otherwise.
-    """
-    try:
-        import platformdirs
-    except ModuleNotFoundError as err:
-        log.error(
-            "Missing platformdirs. Re-install sdmx with pip install sdmx1[validation]"
-        )
-        raise err
-
-    # If the user has no preference, get the schemas from the local cache directory
-    if not schema_dir:
-        schema_dir = platformdirs.user_cache_path("sdmx")
-
-    msg_doc = etree.parse(msg)
-
-    # Make sure the message is a supported type
-    supported_elements = [
-        "CodelistQuery",
-        "DataStructureQuery",
-        "GenericData",
-        "GenericMetadata",
-        "GenericTimeSeriesData",
-        "MetadataStructureQuery",
-        "Structure",
-        "StructureSpecificData",
-        "StructureSpecificMetadata",
-        "StructureSpecificTimeSeriesData",
-    ]
-    root_elem_name = msg_doc.docinfo.root_name
-    if root_elem_name not in supported_elements:
-        raise NotImplementedError
-
-    message_xsd = schema_dir.joinpath("SDMXMessage.xsd")
-    if not message_xsd.exists():
-        raise ValueError
-
-    # Turn the XSD into a schema object
-    xml_schema_doc = etree.parse(message_xsd)
-    xml_schema = etree.XMLSchema(xml_schema_doc)
-
-    return xml_schema.validate(msg_doc)
-    # return xml_schema.assertValid(msg_doc)
-
-
-def install_schemas(schema_dir: Optional[Path] = None) -> None:
-    """Cache XML Schema documents locally for use during message validation.
-
-    Parameters
-    ----------
-    schema_dir
-        The directory where XSD schemas will be downloaded to.
-    """
-    import io
-    import zipfile
-
-    import platformdirs
-    import requests
-
-    # If the user has no preference, download the schemas to the local cache directory
-    if not schema_dir:
-        schema_dir = platformdirs.user_cache_path("sdmx")
-    schema_dir.mkdir(exist_ok=True, parents=True)
-
-    # Check the latest release to get the URL to the schema zip
-    release_url = "https://api.github.com/repos/sdmx-twg/sdmx-ml-v2_1/releases/latest"
-    gh_headers = {
-        "Accept": "application/vnd.github+json",
-        "X-GitHub-Api-Version": "2022-11-28"
-    }
-    resp = requests.get(url=release_url, headers=gh_headers)
-    zipball_url = resp.json().get("zipball_url")
-
-    # Download the zipped content and find the schemas within
-    resp = requests.get(url=zipball_url, headers=gh_headers)
-    zipped = zipfile.ZipFile(io.BytesIO(resp.content))
-    schemas = [n for n in zipped.namelist() if "schemas" in n and n.endswith(".xsd")]
-
-    # Extract the schemas to the destination directory
-    # We can't use ZipFile.extract here because it will keep the directory structure
-    for xsd in schemas:
-        xsd_path = zipfile.Path(zipped, at=xsd)
-        target = schema_dir.joinpath(xsd_path.name)
-        target.write_text(xsd_path.read_text())
 
 
 def reference(obj, parent=None, tag=None, style=None):
@@ -162,45 +62,51 @@ def reference(obj, parent=None, tag=None, style=None):
 
     elem = Element(tag)
 
-    if isinstance(obj, model.MaintainableArtefact):
-        ma = obj
-    else:
-        try:
-            # Get the ItemScheme for an Item
-            parent = parent or obj.get_scheme()
-        except AttributeError:  # pragma: no cover
-            # No `parent` and `obj` is not an Item with a .get_scheme() method
-            # NB this does not occur in the test suite
-            pass
-
-        if not parent:
-            raise NotImplementedError(
-                f"Cannot write reference to {repr(obj)} without parent"
-            )
-
-        ma = parent
-
+    # assert style
     if style == "URN":
         ref = Element(":URN", obj.urn)
     elif style == "Ref":
-        args = {
-            "agencyID": getattr(ma.maintainer, "id", None),
-            "id": obj.id,
-            "maintainableParentID": ma.id if parent else None,
-            "maintainableParentVersion": ma.version if parent else None,
-            "version": ma.version,
-            "package": model.PACKAGE[ma.__class__.__name__],
-        }
-        for candidate in (obj.__class__, getattr(ma.__class__, "_Item", None)):
+        # Element attributes
+        attrib = dict(id=obj.id)
+
+        # Identify a maintainable artifact; either `obj` or its `parent`
+        if isinstance(obj, model.MaintainableArtefact):
+            ma = obj
+            attrib.update(version=obj.version)
+        else:
             try:
-                args["class"] = etree.QName(tag_for_class(candidate)).localname
+                # Get the ItemScheme for an Item
+                parent = parent or obj.get_scheme()
+            except AttributeError:  # pragma: no cover
+                # No `parent` and `obj` is not an Item with a .get_scheme() method
+                # NB this does not occur in the test suite
+                pass
+
+            if not parent:
+                raise NotImplementedError(
+                    f"Cannot write reference to {obj!r} without parent"
+                )
+
+            ma = parent
+            attrib.update(
+                maintainableParentVersion=ma.version,
+                maintainableParentID=ma.id,
+            )
+
+        attrib.update(
+            agencyID=getattr(ma.maintainer, "id", None),
+            package=model.PACKAGE[type(ma).__name__],
+        )
+
+        # "class" attribute: either the type of `obj`, or the item type of an ItemScheme
+        for candidate in (type(obj), getattr(type(ma), "_Item", None)):
+            try:
+                attrib["class"] = etree.QName(tag_for_class(candidate)).localname
                 break
             except ValueError:
                 pass
 
-        ref = Element(":Ref", **args)
-    else:  # pragma: no cover
-        raise ValueError(style)
+        ref = Element(":Ref", **attrib)
 
     elem.append(ref)
     return elem
@@ -233,7 +139,7 @@ def _dm(obj: message.DataMessage):
             attrib["structureID"] = ds.structured_by.id
 
             # Reference by URN if possible, otherwise with a <Ref> tag
-            style = "URN" if ds.structured_by.urn else "Ref"
+            style: RefStyle = "URN" if ds.structured_by.urn else "Ref"
             dsd_ref = reference(ds.structured_by, tag="com:Structure", style=style)
 
             if isinstance(obj.observation_dimension, model.DimensionComponent):
@@ -285,7 +191,8 @@ def _sm(obj: message.StructureMessage):
         container = Element(f"str:{tag}")
         for s in filter(lambda s: not s.is_external_reference, coll.values()):
             container.append(writer.recurse(s))
-        structures.append(container)
+        if len(container):
+            structures.append(container)
 
     if obj.footer:
         elem.append(writer.recurse(obj.footer))
@@ -364,29 +271,31 @@ def _a(obj: model.Annotation):
     elem = Element("com:Annotation")
     if obj.id:
         elem.attrib["id"] = obj.id
+    if obj.title:
+        elem.append(Element("com:AnnotationTitle", obj.title))
     if obj.type:
         elem.append(Element("com:AnnotationType", obj.type))
     elem.extend(i11lstring(obj.text, "com:AnnotationText"))
     return elem
 
 
-def annotable(obj, **kwargs):
-    cls = kwargs.pop("_tag", tag_for_class(obj.__class__))
+def annotable(obj, *args, **kwargs) -> etree._Element:
+    # Determine tag
+    tag = kwargs.pop("_tag", tag_for_class(obj.__class__))
+
+    # Write Annotations
+    e_anno = Element("com:Annotations", *[writer.recurse(a) for a in obj.annotations])
+    if len(e_anno):
+        args = args + (e_anno,)
+
     try:
-        elem = Element(cls, **kwargs)
+        return Element(tag, *args, **kwargs)
     except AttributeError:  # pragma: no cover
-        print(repr(obj), cls, kwargs)
+        print(repr(obj), tag, kwargs)
         raise
 
-    if len(obj.annotations):
-        e_anno = Element("com:Annotations")
-        e_anno.extend(writer.recurse(a) for a in obj.annotations)
-        elem.append(e_anno)
 
-    return elem
-
-
-def identifiable(obj, **kwargs):
+def identifiable(obj, *args, **kwargs) -> etree._Element:
     """Write :class:`.IdentifiableArtefact`.
 
     Unless the keyword argument `_with_urn` is :data:`False`, a URN is generated for
@@ -402,22 +311,25 @@ def identifiable(obj, **kwargs):
             kwargs.setdefault("urn", urn)
     except (AttributeError, ValueError):
         pass
-    return annotable(obj, **kwargs)
+    return annotable(obj, *args, **kwargs)
 
 
-def nameable(obj, **kwargs):
-    elem = identifiable(obj, **kwargs)
-    elem.extend(i11lstring(obj.name, "com:Name"))
-    elem.extend(i11lstring(obj.description, "com:Description"))
-    return elem
+def nameable(obj, *args, **kwargs) -> etree._Element:
+    return identifiable(
+        obj,
+        *i11lstring(obj.name, "com:Name"),
+        *i11lstring(obj.description, "com:Description"),
+        *args,
+        **kwargs,
+    )
 
 
-def maintainable(obj, **kwargs):
+def maintainable(obj, *args, **kwargs) -> etree._Element:
     kwargs.setdefault("version", obj.version)
     kwargs.setdefault("isExternalReference", str(obj.is_external_reference).lower())
     kwargs.setdefault("isFinal", str(obj.is_final).lower())
     kwargs.setdefault("agencyID", getattr(obj.maintainer, "id", None))
-    return nameable(obj, **kwargs)
+    return nameable(obj, *args, **kwargs)
 
 
 # §3.5: Item Scheme
@@ -502,34 +414,41 @@ def _contact(obj: model.Contact):
 
 
 @writer
-def _component(obj: model.Component):
-    elem = identifiable(obj)
-    if obj.concept_identity:
-        elem.append(
+def _component(obj: model.Component, dsd):
+    child = []
+    attrib = dict()
+
+    try:
+        child.append(
             reference(obj.concept_identity, tag="str:ConceptIdentity", style="Ref")
         )
-    if obj.local_representation:
-        elem.append(
+    except AttributeError:  # pragma: no cover
+        pass  # concept_identity is None
+
+    try:
+        child.append(
             writer.recurse(obj.local_representation, "LocalRepresentation", style="Ref")
         )
-    # DataAttribute only
-    try:
-        elem.append(writer.recurse(cast(model.DataAttribute, obj).related_to))
-    except AttributeError:
-        pass
-    except NotImplementedError:  # pragma: no cover
-        if getattr(obj, "related_to", None) is None:
-            pass  # .related_to not set
-        else:
-            raise  # Some other NotImplementedError
+    except NotImplementedError:
+        pass  # None
 
-    return elem
+    if isinstance(obj, model.DataAttribute) and obj.usage_status:
+        child.append(writer.recurse(obj.related_to, dsd))
+
+        # assignmentStatus attribute
+        if obj.usage_status:
+            attrib["assignmentStatus"] = obj.usage_status.name.title()
+    elif isinstance(obj, model.Dimension):
+        # position attribute
+        attrib["position"] = str(obj.order)
+
+    return identifiable(obj, *child, **attrib)
 
 
 @writer
-def _cl(obj: model.ComponentList):
+def _cl(obj: model.ComponentList, *args):
     elem = identifiable(obj)
-    elem.extend(writer.recurse(c) for c in obj.components)
+    elem.extend(writer.recurse(c, *args) for c in obj.components)
     return elem
 
 
@@ -569,7 +488,12 @@ def _dks(obj: model.DataKeySet):
 
 @writer
 def _ms(obj: model.MemberSelection):
-    elem = Element("com:KeyValue", id=obj.values_for.id)
+    tag = {
+        model.Dimension: "KeyValue",
+        model.DataAttribute: "Attribute",
+    }[type(obj.values_for)]
+
+    elem = Element(f"com:{tag}", id=obj.values_for.id)
     elem.extend(
         # cast(): as of PR#30, only MemberValue is supported here
         Element("com:Value", cast(model.MemberValue, mv).value)
@@ -611,22 +535,26 @@ def _cc(obj: model.ContentConstraint):
 
 
 @writer
-def _nsr(obj: model.NoSpecifiedRelationship):
+def _nsr(obj: model.NoSpecifiedRelationship, *args):
     elem = Element("str:AttributeRelationship")
     elem.append(Element("str:None"))
     return elem
 
 
 @writer
-def _pmr(obj: model.PrimaryMeasureRelationship):
+def _pmr(obj: model.PrimaryMeasureRelationship, dsd: model.DataStructureDefinition):
     elem = Element("str:AttributeRelationship")
     elem.append(Element("str:PrimaryMeasure"))
-    elem[-1].append(Element(":Ref", id="(not implemented)"))
+    try:
+        ref_id = dsd.measures[0].id
+    except IndexError:
+        ref_id = "(not implemented)"  # MeasureDescriptor is empty
+    elem[-1].append(Element(":Ref", id=ref_id))
     return elem
 
 
 @writer
-def _dr(obj: common.DimensionRelationship):
+def _dr(obj: common.DimensionRelationship, *args):
     elem = Element("str:AttributeRelationship")
     for dim in obj.dimensions:
         elem.append(Element("str:Dimension"))
@@ -635,7 +563,7 @@ def _dr(obj: common.DimensionRelationship):
 
 
 @writer
-def _gr(obj: common.GroupRelationship):
+def _gr(obj: common.GroupRelationship, *args):
     elem = Element("str:AttributeRelationship")
     elem.append(Element("str:Group"))
     elem[-1].append(Element(":Ref", id=getattr(obj.group_key, "id", None)))
@@ -658,11 +586,11 @@ def _dsd(obj: model.DataStructureDefinition):
     elem.append(Element("str:DataStructureComponents"))
 
     # Write in a specific order
-    elem[-1].append(writer.recurse(obj.dimensions))
+    elem[-1].append(writer.recurse(obj.dimensions, None))
     for group in obj.group_dimensions.values():
         elem[-1].append(writer.recurse(group))
-    elem[-1].append(writer.recurse(obj.attributes))
-    elem[-1].append(writer.recurse(obj.measures))
+    elem[-1].append(writer.recurse(obj.attributes, obj))
+    elem[-1].append(writer.recurse(obj.measures, None))
 
     return elem